+++ src/main/java/org/opensearch/knn/index/codec/KNN1040Codec/KNN1040ScalarQuantizedUtils.java
@@ -6,8 +6,8 @@
 package org.opensearch.knn.index.codec.KNN1040Codec;
 
 import lombok.experimental.UtilityClass;
-import org.apache.lucene.codecs.lucene104.QuantizedByteVectorValues;
 import org.apache.lucene.index.KnnVectorValues;
+import org.apache.lucene.util.quantization.QuantizedByteVectorValues;
 
 import java.io.IOException;
 import java.lang.reflect.Field;